requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.2
django-allauth==0.35.0
django-anymail==3.0
django-appconf==1.0.2
django-countries==5.3.1
django-select2==6.1.0
django-simple-history==2.2.0
django-taggit==0.22.2
gunicorn==19.9.0
idna==2.6
oauthlib==2.0.6
psycopg2-binary==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
static3==0.7.0
urllib3==1.22
whitenoise==3.3.1
